feat(billing.history): revamp history display
ref: MANAGER-4148
