+++ packages/manager/apps/dedicated/client/app/billing/main/history/debt/details/billing-main-history-debt-details.html
@@ -20,6 +20,7 @@
         >
             <span
                 data-translate="{{ :: 'billing_history_details_type_' + $row.type.toLowerCase() }}"
+                data-translate-default="-"
             ></span>
         </oui-column>
         <oui-column
