+++ packages/manager/apps/dedicated/client/app/billing/main/history/history.module.js
@@ -1,16 +1,37 @@
 import angular from 'angular';
 import 'angular-translate';
+import 'angular-ui-bootstrap';
+import 'ovh-api-services';
 import 'ovh-ui-angular';
+import '@uirouter/angularjs';
+import ngAtInternet from '@ovh-ux/ng-at-internet';
+import ngOvhApiWrappers from '@ovh-ux/ng-ovh-api-wrappers';
+import ngOvhPaymentMethod from '@ovh-ux/ng-ovh-payment-method';
+import ovhManagerCore from '@ovh-ux/manager-core';
 
 import component from './history.component';
 import routing from './billing-main-history.routes';
 
+import legacyHistory from './legacy/history.module';
+
 const moduleName = 'ovhManagerDedicatedBillingHistory';
 
 angular
-  .module(moduleName, ['oui', 'pascalprecht.translate'])
+  .module(moduleName, [
+    legacyHistory,
+    ngAtInternet,
+    ngOvhApiWrappers,
+    ngOvhPaymentMethod,
+    'oui',
+    'ovh-api-services',
+    ovhManagerCore,
+    'pascalprecht.translate',
+    'ui.bootstrap',
+    'ui.router',
+  ])
   .config(routing)
   .component('billingHistory', component)
-  .run(/* @ngTranslationsInject:json ./translations */);
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(/* @ngTranslationsInject:json ./postalMailOptions/translations */);
 
 export default moduleName;
