+++ packages/manager/apps/dedicated/client/app/billing/main/history/legacy/history.component.js
@@ -1,0 +1,12 @@
+import controller from './billing-main-history.controller';
+import template from './billing-main-history.html';
+
+export default {
+  bindings: {
+    currentUser: '<',
+    filters: '<',
+    onListParamsChange: '<',
+  },
+  controller,
+  template,
+};
