+++ packages/manager/apps/dedicated/client/app/billing/main/history/legacy/history.module.js
@@ -1,0 +1,14 @@
+import angular from 'angular';
+import 'angular-translate';
+import 'ovh-ui-angular';
+
+import component from './history.component';
+
+const moduleName = 'ovhManagerDedicatedBillingLegacyHistory';
+
+angular
+  .module(moduleName, ['oui', 'pascalprecht.translate'])
+  .component('billingLegacyHistory', component)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
