+++ packages/manager/apps/dedicated/client/app/billing/main/history/legacy/translations/Messages_fr_FR.json
@@ -1,0 +1,29 @@
+{
+  "billing_main_history_legacy_title": "Mes factures",
+  "billing_main_history_legacy_loading_errors": "Une erreur est survenue lors de la récupération de vos factures.",
+  "billing_main_history_legacy_export_error": "Une erreur est survenue lors de l'exportation de vos factures au format csv.",
+  "billing_main_history_legacy_invoices_postal_mail": "Recevoir mes factures au format papier",
+  "billing_main_history_legacy_table_actions": "Actions",
+  "billing_main_history_legacy_table_actions_title": "Mes factures : actions",
+  "billing_main_history_legacy_table_action_pay_debt": "Régler mon solde immédiatement",
+  "billing_main_history_legacy_table_action_pay_debt_title": "Mes factures : régler mon solde immédiatement",
+  "billing_main_history_legacy_table_action_export": "Exporter en CSV",
+  "billing_main_history_legacy_table_action_export_title": "Mes factures : exporter en CSV",
+  "billing_main_history_legacy_table_id": "Référence",
+  "billing_main_history_legacy_table_order_id": "Numéro de commande",
+  "billing_main_history_legacy_table_date": "Date d'émission",
+  "billing_main_history_legacy_table_total": "Montant",
+  "billing_main_history_legacy_table_total_with_VAT": "Montant TTC",
+  "billing_main_history_legacy_table_balance_due_amount": "Solde à payer",
+  "billing_main_history_legacy_table_balance_due_date": "Date due",
+  "billing_main_history_legacy_table_unavailable": "Information non-disponible",
+  "billing_main_history_legacy_table_immediately": "Immédiate",
+  "billing_main_history_legacy_table_debt_paid": "Payée",
+  "billing_main_history_legacy_table_download_html": "Voir la version HTML",
+  "billing_main_history_legacy_table_download_pdf": "Voir la version PDF",
+  "billing_main_history_legacy_table_download_pdf_beta": "Voir la version PDF (Beta)",
+  "billing_main_history_legacy_table_debt_detail": "Voir les détails pour cette dette",
+  "billing_main_history_legacy_table_debt_pay": "Régler mon solde immédiatement",
+  "billing_main_history_legacy_debt_to_pay": "Un montant de <strong>{{t0}}</strong> est dû pour le renouvellement de vos services.",
+  "billing_main_history_legacy_debt_automatically_collected": "Un montant de <strong>{{t0}}</strong> sera prélevé de votre moyen de paiement par défaut."
+}
