+++ packages/manager/apps/dedicated/client/app/billing/main/history/translations/Messages_fr_CA.json
@@ -1,5 +1,4 @@
 {
-  "billing_main_history_title": "Mes factures",
   "billing_main_history_loading_errors": "Une erreur est survenue lors de la récupération de vos factures.",
   "billing_main_history_export_error": "Une erreur est survenue lors de l'exportation de vos factures au format csv.",
   "billing_main_history_invoices_postal_mail": "Recevoir mes factures au format papier",
