+++ packages/manager/apps/dedicated/client/app/billing/main/history/translations/Messages_fr_FR.json
@@ -1,29 +1,40 @@
 {
-  "billing_main_history_title": "Mes factures",
   "billing_main_history_loading_errors": "Une erreur est survenue lors de la récupération de vos factures.",
   "billing_main_history_export_error": "Une erreur est survenue lors de l'exportation de vos factures au format csv.",
   "billing_main_history_invoices_postal_mail": "Recevoir mes factures au format papier",
-  "billing_main_history_table_actions": "Actions",
+  "billing_main_history_table_actions": "Actions de masse",
   "billing_main_history_table_actions_title": "Mes factures : actions",
   "billing_main_history_table_action_pay_debt": "Régler mon solde immédiatement",
   "billing_main_history_table_action_pay_debt_title": "Mes factures : régler mon solde immédiatement",
-  "billing_main_history_table_action_export": "Exporter en CSV",
-  "billing_main_history_table_action_export_title": "Mes factures : exporter en CSV",
+  "billing_main_history_table_action_export_selection": "Exporter la sélection au format CSV",
+  "billing_main_history_table_action_export_selection_pdf": "Exporter la séléction au format PDF",
+  "billing_main_history_table_action_export_all": "Tout exporter en CSV",
+  "billing_main_history_table_action_export_all_pdf": "Tout exporter en PDF",
   "billing_main_history_table_id": "Référence",
+  "billing_main_history_table_category": "Catégorie",
+  "billing_main_history_table_category_autorenew": "Renouvellement automatique",
+  "billing_main_history_table_category_purchase-cloud": "Cloud",
+  "billing_main_history_table_category_purchase-servers": "Server",
+  "billing_main_history_table_category_purchase-telecom": "Telecom",
+  "billing_main_history_table_category_purchase-web": "Web hosting",
   "billing_main_history_table_order_id": "Numéro de commande",
   "billing_main_history_table_date": "Date d'émission",
-  "billing_main_history_table_total": "Montant",
+  "billing_main_history_table_total": "Montant HT",
   "billing_main_history_table_total_with_VAT": "Montant TTC",
-  "billing_main_history_table_balance_due_amount": "Solde à payer",
-  "billing_main_history_table_balance_due_date": "Date due",
-  "billing_main_history_table_unavailable": "Information non-disponible",
-  "billing_main_history_table_immediately": "Immédiate",
-  "billing_main_history_table_debt_paid": "Payée",
+  "billing_main_history_table_balance_status": "Statut",
+  "billing_main_history_table_balance_due_amount": "Solde",
+  "billing_main_history_table_balance_status_UNPAID": "Impayée",
+  "billing_main_history_table_balance_status_PAID": "Payée",
+  "billing_main_history_table_balance_status_REFUNDED": "Remboursée",
+  "billing_main_history_table_balance_status_TO_BE_PAID": "À régler",
+  "billing_main_history_table_balance_status_UNMATURED": "À régler",
   "billing_main_history_table_download_html": "Voir la version HTML",
   "billing_main_history_table_download_pdf": "Voir la version PDF",
   "billing_main_history_table_download_pdf_beta": "Voir la version PDF (Beta)",
   "billing_main_history_table_debt_detail": "Voir les détails pour cette dette",
   "billing_main_history_table_debt_pay": "Régler mon solde immédiatement",
   "billing_main_history_debt_to_pay": "Un montant de <strong>{{t0}}</strong> est dû pour le renouvellement de vos services.",
-  "billing_main_history_debt_automatically_collected": "Un montant de <strong>{{t0}}</strong> sera prélevé de votre moyen de paiement par défaut."
+  "billing_main_history_debt_automatically_collected": "Un montant de <strong>{{t0}}</strong> sera prélevé de votre moyen de paiement par défaut.",
+  "billing_main_history_table_export_success": "Nous allons vous envoyer un email contenant un lien de téléchargement vous permettant de récupérer votre export de factures. En fonction du nombre de factures, cela peut prendre quelques minutes.",
+  "billing_main_history_table_export_error": "Une erreur est survenue lors de la demande d'export de vos factures. {{ message }}"
 }
