+++ packages/manager/apps/dedicated/client/app/billing/main/translations/Messages_fr_FR.json
@@ -1,5 +1,5 @@
 {
-  "billing_main_title": "Facturation",
+  "billing_main_title": "Mes factures",
   "billing_main_pay_as_you_go_tab": "Pay As You Go",
   "billing_main_bills_tab": "Factures",
   "billing_main_payments_tab": "Suivi des paiements"
