+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_de_DE.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(früher OVH Manager)",
   "menu_fidelity": "Mein Treuebereich",
   "menu_vouchers": "Meine Gutscheine",
-  "menu_billing": "Abrechnung",
   "menu_bills": "Meine Rechnungen",
   "menu_support": "Meine Support-Anfragen",
   "menu_payment_methods": "Zahlungsarten",
