+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_es_US.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(antigua cuenta de OVH)",
   "menu_fidelity": "Área de fidelidad",
   "menu_vouchers": "Mis códigos promocionales",
-  "menu_billing": "Facturación",
   "menu_bills": "Mis facturas",
   "menu_payment_methods": "Moyens de paiement",
   "menu_subscriptions": "Mis suscripciones",
