+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_fr_FR.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(anciennement mon compte OVH)",
   "menu_fidelity": "Mon espace fidélité",
   "menu_vouchers": "Mes vouchers",
-  "menu_billing": "Facturation",
   "menu_bills": "Mes factures",
   "menu_support": "Mes demandes d'assistance",
   "menu_payment_methods": "Moyens de paiement",
