+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_it_IT.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(precedentemente conto OVH)",
   "menu_fidelity": "Il tuo spazio fedeltà",
   "menu_vouchers": "I tuoi voucher",
-  "menu_billing": "Fatturazione",
   "menu_bills": "Le tue fatture",
   "menu_support": "Le tue richieste di assistenza",
   "menu_payment_methods": "Modalità di pagamento",
