+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_lt_LT.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(formerly known as my OVH account)",
   "menu_fidelity": "My Loyalty control panel",
   "menu_vouchers": "My vouchers",
-  "menu_billing": "Billing",
   "menu_bills": "My bills",
   "menu_support": "My support tickets ",
   "menu_payment_methods": "Payment method",
