+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_pl_PL.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(dawniej konto OVH)",
   "menu_fidelity": "Konto prepaid",
   "menu_vouchers": "Moje kody",
-  "menu_billing": "Płatności",
   "menu_bills": "Faktury",
   "menu_support": "Moje zgłoszenia",
   "menu_payment_methods": "Sposoby płatności",
