+++ packages/manager/apps/dedicated/client/app/components/sidebar-menu/account/translations/Messages_pt_PT.json
@@ -17,7 +17,6 @@
   "menu_ovhaccount_old_name": "(antiga conta OVH)",
   "menu_fidelity": "Espaço Fidelização (pontos)",
   "menu_vouchers": "Vouchers",
-  "menu_billing": "Faturação",
   "menu_bills": "Faturas",
   "menu_support": "Os meus pedidos de assistência",
   "menu_payment_methods": "Métodos de pagamento",
