+++ packages/manager/apps/dedicated/package.json
@@ -34,6 +34,7 @@
     "@ovh-ux/manager-models": "^1.0.0",
     "@ovh-ux/manager-nasha": "^1.0.0",
     "@ovh-ux/manager-navbar": "^3.0.0",
+    "@ovh-ux/manager-ng-layout-helpers": "^1.0.1",
     "@ovh-ux/manager-preloader": "^1.0.0",
     "@ovh-ux/manager-server-sidebar": "^1.0.0",
     "@ovh-ux/manager-support": "^0.6.0",
