+++ packages/manager/modules/models/index.js
@@ -1,10 +1,30 @@
+import Bill from './billing/Bill.class';
 import BillingService from './BillingService/BillingService.class';
 import Certificate from './certificate/Certificate.class';
+import DebtAccount from './billing/DebtAccount.class';
 import Price from './Price/Price.class';
 import Service from './service/Service.class';
 import Ticket from './support/ticket/ticket.class';
 import User from './user/User.class';
 
-export { BillingService, Certificate, Price, Service, Ticket, User };
+export {
+  Bill,
+  BillingService,
+  Certificate,
+  DebtAccount,
+  Price,
+  Service,
+  Ticket,
+  User,
+};
 
-export default { BillingService, Certificate, Price, Service, Ticket, User };
+export default {
+  Bill,
+  BillingService,
+  Certificate,
+  DebtAccount,
+  Price,
+  Service,
+  Ticket,
+  User,
+};
