+++ packages/manager/modules/models/user/User.class.js
@@ -23,4 +23,12 @@ export default class User {
   get isEnterprise() {
     return some(this.certificates, (certificate) => certificate.isEnterprise());
   }
+
+  get isVATNeeded() {
+    return !['CA', 'QC', 'WE', 'WS'].includes(this.ovhSubsidiary);
+  }
+
+  canHaveInvoicesByPostalMail() {
+    return this.billingCountry === 'FR' && this.legalform === 'individual';
+  }
 }
